Scrape constituency names
Scrapes the name of constituency for members. Members are displayed grouped by district. The district name is only displayed in the upper most row of each district. When the district name is present, it is stored. When a row is without a district name, the stored district name is used.