Reference tds by negative index
The source table on the 2016 contains one more column than
the 2012 and 2008 term pages. As a result, indexing from
the left is not consistent However, indexing from the
right is (for these three terms).

For example, the party name is always the last field. However
its index from the start varies between term 2016 and terms
2008 and 2012.

https://en.wikipedia.org/wiki/List_of_MPs_elected_in_the_Mongolian_legislative_election,_2016
https://en.wikipedia.org/wiki/List_of_MPs_elected_in_the_Mongolian_legislative_election,_2012
https://en.wikipedia.org/wiki/List_of_MPs_elected_in_the_Mongolian_legislative_election,_2008