Add PartyListKhuralMember
Class added to represent member elected from party list.

The 2012 term differs from the 2008 and 2016 terms because it
contains party list members. The party list members are listed
in a table with a layout that differs from that of the constituency
list members. Previously, the constituency name for each party list
member was being set as that member's personal name.The new class returns nil for this field as party list members do not have a constituency.

The class inherits directly from NokogiriDocument as fields
created with FieldSerializer cannot be inherited currently.