Remove content_class property
A step towards breaking MemberTable into two separate classes: one
for constituency members, the other for party list members.

Since we are treating NokogiriDocument as a gem, we don't want to
change its functionality... the extra property is too specific to
this scraper to become part of the gem.

Instead of injecting the member class we want to create into
NokogiriDocument, we will create to separate table classes for each
type of member.